Avoid dependency on Rust nightly for capturing test output
Instead of using unstable API available only on Rust nightly, this uses
the gag crate to capture stdout and stderr of the test scenarios. This
has one downside though, that crate currently works only on Unix-like
systems. But it is strictly more general as it redirects file
descriptors instead of just change the Rust-specific stdout and stderr
locked readers and writers.